Add toggle option for menu on mobile. NOTE: known issue is present
When a user goes from a large screen to a mobile sized screen and toggles/untoggles the navbar, the menu bar gets messed up when the user goes back to the large screen. large screen -> mobile screen -> toggle menu -> large screen == messed up navbar. Also when the navbar is toggled in mobile sized mode, and the menu is thus hidden, the menu content is also hidden when the user goes back to the large screen